Agent: set client environment based on client configuration info (#6194)
CLOSE https://linear.app/sourcegraph/issue/CODY-3817

1. set shell capability based on client configuration.

This change updates the `vscode-shim` to set the `shell` capability
based on the client's reported capabilities. If the client set the
`shell` environment variable to `enabled`, it will be set to `'agent'`,
otherwise it is set to `undefined`.

When enabled, the client must be running with node integration enabled
that supports `child_process.spawn`. This is needed for the agent to
spawn a shell process to run commands in IDEs that support node runtime.

This ensures that the shell command functionality is only enabled when
the client has the necessary support.

2. set client environment based on configuration

This change updates the `vscode-shim` to set the client environment
variables based on the client's reported capabilities. The `appName` and
`shell` environment variables are now set dynamically using the
`setClientEnv` function, which is called when the client information is
updated.

This ensures that the environment variables are properly configured to
match the client's capabilities, enabling features like the agent shell
process spawning when the `shell` capability is enabled.

### Notes

Currently, no clients have enabled this feature yet.

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->


![image](https://github.com/user-attachments/assets/49e358ce-1a72-4ca4-af3d-85eb9c099b66)

1. Build JB / VS with agent built with this branch and set
`clientInfo?.capabilities?.shell` to `enabled`
2. Only Deep Cody is using the shell context right now, so you can ask
Deep Cody a question and see if it is able to use shell context to
answer your question. Something like "how many files are in the root of
this codebase"

### Before

Before in JB: 

> Cody by Sourcegraph: █ CodyTool executing 1 commands...:
> window.showErrorMessage: Shell command is not supported in your
current workspace.


![image](https://github.com/user-attachments/assets/ae0b686f-7deb-4bb7-aaf6-ea6ca7a8792c)


After


![image](https://github.com/user-attachments/assets/cdd20a67-4f5a-4432-8b12-9495f36477c0)


![image](https://github.com/user-attachments/assets/a0320f0c-d3b0-4f48-86ef-53b9b16fb38f)


## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->

Agent: Add shell capability to enable agent shell process spawning.