feat(webviews): add telemetry events and session-based dismissal for notices (#6266)
Follow up on https://github.com/sourcegraph/cody/pull/6245
CLOSE
https://linear.app/sourcegraph/issue/CODY-4346/implement-the-eligible-for-teams-upgrade-triggering
CLOSE
https://linear.app/sourcegraph/issue/CODY-4345/add-telemetry-to-the-eligible-for-teams-upgrade-cta

This PR introduces session-based dismissal for notices in the Cody
extension. The 'DogfoodS2' notice continued to be dismissed only for the
current session, allowing them to be shown again if the user logs out
and logs back in. The 'TeamsUpgrade' notice is now dismissed forever
after the user has clicked on the dismiss button once.

The changes include:
- Adding a `useLocalStorage` hook to manage dismissed notices in local
storage
- Introducing a `sessionDismissedNotices` state to track notices
dismissed for the current session
- Updating the `dismissNotice` function to handle both permanent and
session-based dismissal
- Modifying the `activeNotice` useMemo to handle the new session-based
dismissal logic for the 'DogfoodS2' notice
- Record events on button clicks

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

- Verify that the 'DogfoodS2' notice is shown when the extension is
first loaded
- Dismiss the 'DogfoodS2' notice and verify that it does not reappear
until the user logs out and logs back in
- Verify that other notices are still dismissed permanently as expected
- 
## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->