fix line level aggregation logic for recent edits (#6282)
## Context
We aggregate the changes which are made on the same lines together for
recent edits diff calculation. However, if we replace the text block in
the vscode editor, the vscode editor create seperate events for each
line even though the change is a single replacement. In the PR we group
the changes to if the timestamps are same 2 different change even if
they are added on different lines.

## Test plan
Added test cases