NLS search for OneBox (#6263)
Currently for search intent, simply the context items fetched using the
`getCodyContext` API (which we use to pass to the LLM) were shown as
search results.

This PR uses the graphql search api with `patternType=NLS` to fetch the
search results for the query and display them in the assistant response.

This PR also converts the file, repo & directory mentions to search
filters and ignores the rest mentions for search intent.

---------

Co-authored-by: Camden Cheek <camden@ccheek.com>