feat(deep cody): show model on pre-release versions only
This PR adds a new function `isClientOnStableVersion` to the `clientCapabilities` module. This function checks if the client is on a stable version of the extension by parsing the version string and checking if the minor version is even.

This is used in the `syncModels` function to only display the Deep Cody feature in pre-release versions of the extension.

## Changelog

- Added `isClientOnStableVersion` function to `clientCapabilities` module