feat(audoedit): inline renderer POC (#6214)
- Implements an experimental autoedits inline render POC. It shows
autoedit suggestions using the inline completion item provider (used for
autocomplete) and line decorations.
- I plan to improve this initial version with multiple smaller PR
addressing one use case at a time to make the review easier. The main
goal of the POC was to verify the number of cases we can cover with this
approach, uncover unknown blockers, and show how it looks in practice.