map mentions to format expected by intent API (#6321)
Once [this](https://github.com/sourcegraph/cody/pull/6321) lands, we
will be using the new intent model on S2. The new intent model expects
mentions to be mapped to a placeholder, so for example
`@sourcegraph/conc` will be mapped to the text `@repo`. We can't do that
on the backend because we lose information about the type of an
@-mention when we stringify it, so this PR updates the intent code to do
those replacements before making the intent API call.