fix(deep-cody): add experimental model tag
This PR adds the `ModelTag.Experimental` tag to the `getServerModelTags` function which is the cause of Deep Cody currently missing the experimental tag in pre-release. This tag will be used to identify models that are in an experimental phase, which can be useful for displaying additional information or handling these models differently.

Also update the CTA for Deep Cody and add an info section about that daily usage limit.

## Test plan

No additional testing required, as this is a small change to the model tagging logic.