cherry-pick: Prep for the A/B test on Claude 3.5 Haiku as the default chat model for Free users (CODY-4455) (#6304)
This PR does [Run experiment on Claude 3.5 Haiku as the default chat
model for Free
users](https://linear.app/sourcegraph/issue/CODY-4455/run-experiment-on-claude-35-haiku-as-the-default-chat-model-for-free).
This is part of [LLM cost reduction for PLG
users](https://linear.app/sourcegraph/project/llm-cost-reduction-for-plg-users-57f5801d7ea1/issues).

Problem: We want to reduce the LLM cost for PLG users.

This PR prepares for A/B test on Claude 3.5 Haiku as a default chat
model Free users.
Free users with cody-chat-default-to-claude-3-5-haiku feature flag set
to true will have Claude 3.5 Haiku as the default model.
The success criteria are defined in the linear ticket above.

Things the PR does:
- `FeatureFlagProvider.ts` and `sync.ts`: Add the new feature flag and
the main logic.
- `sync.test.ts`: Add tests.

Added tests in `sync.test.ts`

Test this manually
- Go to
https://sourcegraph.com/site-admin/feature-flags/configuration/cody-chat-default-to-claude-3-5-haiku
and add an override for your user to enable the feature flag.
- Build the extension from the branch.
- Singin with the DotCom account.
- You should see Claude 3.5 Haiku as a default Edit model.
<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->

---------

Co-authored-by: Beatrix <68532117+abeatrix@users.noreply.github.com>
(cherry picked from commit 18210bc91b3882cf7d17555404697f0705659e99)