feat(webviews): add Sourcegraph Teams providers
This PR adds new Sourcegraph Teams providers for repositories and directories to the OpenContext menu. These providers are only shown for DotCom users who have the Sourcegraph Teams upgrade feature flag enabled.

The new providers are added with the following URIs:
- `TEAMS_REPOSITORY_PROVIDER_URI`
- `TEAMS_DIRECTORY_PROVIDER_URI`

Additionally, a new `isSourcegraphTeamsProvider` function is added to the `lib/shared/src/context/openctx/api.ts` file to check if a given provider URI belongs to the Sourcegraph Teams providers.

When a user clicks on a Sourcegraph Teams provider in the OpenContext menu, the corresponding Sourcegraph Teams page will be opened in a new tab.

## Test plan

Manual testing.

## Changelog

- Added new Sourcegraph Teams providers for repositories and directories
- Implemented logic to check if a provider is a Sourcegraph Teams provider
- Opened Sourcegraph Teams page in a new tab when a Sourcegraph Teams provider is clicked