feat(audoedit): inline renderer – prevent flickering on acceptance (#6360)
- Fixes the issue with the inline renderer acceptance logic that caused
parts of the suggested text to briefly appear twice in the document,
resulting in a flicker. Initially, I tried an approach where
non-inline-completion-item changes were applied after accepting the
inline completion item (using `InlineCompletionItem['command']`).
However, this made it impossible to keep the undo/redo stack atomic
because the auto-edit suggestion changes were split into two steps:
first, the inline completion item acceptance, and then the additional
changes. I reverted to a solution where we programmatically discard the
inline completion suggestion and use the edit builder to replace the
entire block in one go.
- I removed the `adjustPredictionIfInlineCompletionPossible` call from
the inline manager logic for now because I noticed several suggestions
where new lines modification was not necessary to render inline
completion item because of the update algorithm for inferring the
`insertText`. @hitesh-1997 if there are critical examples where it's
necessary to keep it. Maybe we should split it into several smaller
operations and apply only part of them to the prediction.