feat: prompts library (#6346)
For CODY-3525

This PR turns the action row with "Create prompt" and "Explore prompt
library" into a dropdown with more filters (also tagging).

Tag filters are only available if the backing sourcegraph instance is at
least v5.11 (to be released on Wednesday) or an insider build.

The height of the tags filters is currently limited to 200px, and
becomes scrollable afterwards. We may need a bit more time to understand
how pagination should be used for tags in the editor (or if that should
be used at all), and for now we're just requesting the first 999 tags.
Should be plenty for most customers.


https://github.com/user-attachments/assets/0175aef7-2bed-442e-bd6a-11953bb4d6fb

## Test plan

Manual testing

## Changelog

feat: prompts library