fix: Prevent style leaks in cody web (#6427)
This was an interesting issue to debug: In the Sourcegraph web app we
have a CSS variable `--input-height`. We noticed that when cody was
open, this variable was set to a different value. But nowhere in the
Cody codebase to we set this value explictly.

In turns out that `@vscode/webview-ui-toolkit` constructs a stylesheet
programmatically and sets a couple of global CSS variables.

We only use components from this library in two places and they can
easily be replaced by our own components, thus avoiding this problem
altogether.

NOTE: We'll have to cut a new cody web release for this to fix the web
app issue.

## Test plan

Manual testing by running the web demo. I verified that the CSS
variables are not created anymore and inspected the updated UI, which is
slightly different (e.g. hover effects of buttons) but probably
acceptable.