feat(prompt-editor): Add new ProseMirror-based implementation (#6272)
This PR adds a new version of the prompt editor, based on ProseMirror.

## Video demo

https://github.com/user-attachments/assets/a9fb6410-532e-4b0b-a4eb-e5b5916a3aa7

## Implementation notes

### ProseMirror

We use ProseMirror with a custom schema, that only contains paragraph
nodes, text nodes and mention nodes (the mention "chips"). Programmatic
changes to the editor value happen via transactions, which are applied
to the current state and produce a new state. More on that in the next
section.

### Main 'components'

The 'meat' of the input is modeled as an xstate statechart in
`promptInput.ts` and consists of three main pieces:
- The editor itself, whose state gets updated either via user
interaction or programmatically, via events.
- The data fetching for the mentions menu, including which provider is
currently selected.
- The mentions menu itself, specifically its visibility state and which
item is selected.

Initially I implemented each of these completely separately, but
combined them all into a single statechart. While I normally prefer
modularity, all of these parts have intricate connections, and all of
them make the prompt input what it is, so it seemed to be reasonable to
group them together. The nature of the statechart still makes them easy
to reason about IMO.
I chose to implement the whole system as a statechart so that we have a
mostly descriptive way to specify how the input should behave, in a
central place.

Most importantly though, any operation that can change the value of the
input is implemented as a state transition. This has the advantage that
we can see in a single place which operations are available and how they
will affect the input. And we can also easily test them.

`promptInput-react.ts` is a thin wrapper around the implementation that
makes the input and mentions menu data available via hooks and abstracts
from the state machine.

`PromptEditor.tsx` renders the actual UI elements based on the state and
provides the data from the 'outside world'. Specifically this component
provides the implementation for fetching the mentions data.

Auxiliary modules that support the above are:

- `MentionsMenu.tsx` - for rendering the mentions menu. It uses existing
components for rendering the menu items.
- `MentionsView.tsx` - for rendering the mentions 'chips'. I couldn't
use the existing mentions node component because it was strongly
depending on lexical specific data (but maybe that could be abstracted).
- `lexical-interop.tsx` - makes it possible to use the new component as
a drop in replacement (I hope) by converting lexical editor state to and
from prosemirror state, based on my understanding of the lexical state
that I saw in storybook. I didn't spend a lot of time testing this yet,
but I'll iterate on this as needed.

### Styling

I'm not quite sure what current guidelines are around using CSS directly
and using tailwind. I used tailwind classes mostly where I wanted to
replicate the style of the existing input. Happy to eventually clean
that up.

### Differences to the current implementation

- Renders a decoration around the `@mention` in the input
- `@mention`'s need to be explicitly triggered (like in linear). When
you 'leave' and `@mention` it turns into normal text.
- Currently does not memoize results from providers. I first need to
understand better why/when/how we want this. Given how it's implemented
it should be easy to add though.
- Some difference regarding styling and copy for the mentions menu.

### Things that I wanted to do but couldn't (yet)

- I really wanted to provide a better experience for the mentions menu
while fetching the mentions data (e.g. showing a loading indicator). But
given the nature of the API for fetching the data this was not easily
doable. It's something I want to investigate when working on the new
designs for the menu.
- Eventually we want to use the input also in the prompt library for
creating prompts. For this we might want to introduce a new node type
for placeholders. I don't have enough context on that project yet so I
didn't do anything in this direction, but adding support for this
shouldn't be too difficult.
- There is no support for more complex node types like lists or code
blocks. I think we want to add those eventually but I also need more
information on this first.

## Test plan

Manual testing and unit tests. This PR doesn't impact any of the code
that is currently in use, it only adds the new component so that it can
be used in the future.