fix: remove trailing spaces from extracted query (#6432)
When we remove the mentions, for some reason we're trimming spaces from
the start, but not the end. This means our exact match heuristics won't
work as well for these situations because there's a rogue space at the
end of the query.

This just fixes the logic to trim spaces from both the start and end

## Test plan

This query used to give less relevant results because we kept the
trailing space:
![CleanShot 2024-12-20 at 10 53
31@2x](https://github.com/user-attachments/assets/f49246ee-99cf-47b2-835a-34505ab7ac5d)

(cherry picked from commit 78da6bb02e49a76444ef2eecdb425e36184486d1)