Add security considerations to prompt (#6456)
I tested several PoCs with Deep Cody, which included something in the
context (in my testing, a local file) pointing Cody to run something
dangerous or malicious. I was able to execute:
- network requests
- accessing sensitive files (like /etc/passwd)
- creating arbitrary files

These prompt tweaks were initially suggested by Claude with the prompt
generator, but then I tested and tweaked it further. The command
execution is still hit-and-miss but this definitely improves the overall
behavior. We will continue to rely on explicit user consent to enable
the feature and an allowlist for which commands to execute.

## Test plan
I created PoCs and, with the new prompt, they did not execute
successfully anymore.
<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->

---------

Co-authored-by: Beatrix <68532117+abeatrix@users.noreply.github.com>