Add Cody bench command for NLS (#6497)
This PR adds an NLS strategy to Cody bench that uses the
`nlsSearchQuery` GraphQL endpoint. This lets us run NLS evals against
the actual method used to power omnibox. Previously, we were using the
`getCodyContext` endpoint and adjusting it to use `patterntype:nls`. So
it resolves an important piece of tech debt in our evals.

Addresses SPLF-757