feat(nls): add relevant repo boost (#6502)
This adds the search param to boost a user's relevant repos. It is only
added if the server version supports it (min version 6.0.0 or insiders
build).

I did not yet implement the UI to enable/disable boosting, but wanted to
get this out ASAP for dogfooding. Will work on the interactions next.

## Test plan

Manual testing of a local build.