Experimental proof-of-concept tool use in omnibox (#6510)
Implement a proof of concept tool use response handler in omnibox called
"ToolCody".

As yet, this handler doesn't do anything fancy or anything that can't be
done better by an existing handler (such as an base-model with an
`@repo` chip or DeepCody). It has only one tool "read_file". But it
provides some basic UI scaffolding for rendering assistant responses
that involve multiple "text" and "tool use" steps (thereby going beyond
our classic "chat-based" paradigm).

Because this is a proof of concept, it depends directly on the Anthropic
SDK. See the test plan for instructions for how to enable.

## Changes to prod behavior

There should be no changes to prod behavior, provided this feature is
not activated by setting the user setting.