Record telemetry when executing search
This commit adds the last piece for
https://linear.app/sourcegraph/issue/SRCH-1455/one-box-telemetry. Now
when a message is submitted as search we also record a
'cody.chat-question/executed' event.
The intent (detected and chosen) is part of the submission as well.

I'm not sure if this is correct but I chose to pass the passed
'mentions' as context for the telemetry event, since it's from those
mentions that we compute the scope(s) of the search query.