suggestion mode UI (#6523)
With `auto-edits` being a replacement of `autocomplete`. Based on the
[suggestion](https://sourcegraph.slack.com/archives/C07F8LLKE06/p1735327930973269?thread_ts=1735292806.801459&cid=C07F8LLKE06)
The, PR adds a `drop-down` for the users to choose between the
"autocomplete", "auto-edits" and "Off".

The PR deprecates the 2 earlier `configurations`:
1. `cody.autocomplete.enabled`
2. `cody.experimental.autoedits.enabled`

And introduces the `cody.suggestions.mode: 'autocomplete' | 'auto-edits'
| 'off' ` for the users to choose from.

<img width="549" alt="image"
src="https://github.com/user-attachments/assets/0786fab9-9711-41ed-ab09-8d1c41e28023"
/>


## Test plan
Green CI
Also tested the agent test cases with the new config. Agent takes the
default values from `vscode/package.json` which uses `autocomplete` as
default for suggestion mode and it works.