fix: Remove search context chip on resubmission (#6556)
This commit adds logic to remove the search result context chip when the
query is resubmitted.

We can only implement this for the penultimate input because we only
have a reference to the last input. That suffices however
because when editing a message earlier in the thread, all the following
messages are cleared anyway.

Fixes
https://linear.app/sourcegraph/issue/PROD-421/previous-context-remains-in-follow-up-chatbox-after-resubmitting-query

## Test plan

Manually tested locally in vscode. I created a chat with multiple
messages and edited the first, middle and last one. In all cases the
input was cleared.