fix: Hide search result checkboxes instead of disabling them
When a follow up query to a search result was submitted we've disabled
the checkboxes of the previous search result because changing the
selection wouldn't impact the already submitted follow up query (but
also from an implementation perspective we wouldn't actually be able to
target the follow up input).

But because we store which results have been selected in memory only,
rerendering the search results would show all checkboxes again as
selected (albeit disabled).

This might be confusing to the user.

We would have to persist the UI state somewhere/somehow but there is no
good place for this atm. So to (hopefully) reduce the confusion we
simply hide the checkboxes when the user cannot interact with the search
results anymore.