Fix repo name resolver cache miss due to using separate RepoNameResol… (#6570)
We were creating a separate instance of the RepoNameResolver which was
causing cache miss and in return making graphql calls to fetch the repo
name.

RepoNameResolver internally maintains an LRU and is supposed to be a
singelton.

## Test plan

- monitor network to check getRepoName queries are not made everytime
you press @.
-
## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->

(cherry picked from commit 3f908ae4d94862a511c249d214ef97cacb93e57b)