bench/context: Cache repo IDs (#6569)
I noticed that we call the server to retrieve repo IDs for every example
in the eval set when running context evals. This operation takes ~250ms
for me whereas the actual context search is ~1s. By caching these
results, we can speed up the benchmark by 20%, which is especially
noticeable with larger datasets.

## Test plan
Context bench still works and runs faster.