feat(agentic-context): add feature flag for session usage limit  (#6623)
This adds a new feature flag, `AgenticContextSessionLimit,` to limit the
usage of agentic context to the first 5 human messages in a chat
session, which can be used as a fallback if needed.

This change ensures that we have a fallback plan when the agentic
context is causing issues and being used excessively, especially in
long-running chat sessions, to optimize performance when needed.

Record events:
- `cody.agentic-chat.sessionLimit:skipped` when query is less than 3
words
- `cody.agentic-chat.sessionLimit:hit` when `AgenticContextSessionLimit`
is enabled and the chat messages in current session is over 5 human
messages

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

Manual testing to ensure agentic context is only used for the first 5
human messages when the session limit flag is enabled.

When hit:


![image](https://github.com/user-attachments/assets/431ee086-9cca-410f-a830-5a350f70a3e2)