fix(agentic-context): update search tool prompt and examples (#6632)
As discussed with @camdencheek , our search endpoint for context
retrieval doesn't support Sourcegraph-style search query. This PR
modifies the search examples for the "Search tool" to remove the use of
Sourcegraph-style search query.

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

Ask Cody `Write a unit test for the openExternalAuthUrl function` with
agentic context enabled:

Before: It'd run a code search with `openExternalAuthUrl file:*` as the
search query


![image](https://github.com/user-attachments/assets/7cadb601-ec1c-46de-a89a-ae840e283e0b)

After: It'd run a code search with `openExternalAuthUrl` as the search
query


![image](https://github.com/user-attachments/assets/54ee5002-371a-413d-9f1e-563467b722fe)