feat(auto-edit): improve error logging (#6609)
- Most changes are whitespace-related; hide them for easier review.
- Wraps the logic inside the autoedit provider's
`provideInlineCompletionItems` method in a try-catch block and logs
errors to our telemetry backend and Sentry.
- Eases the conditions for filtering out errors on the client by
checking for the `sourcegraph.cody-ai` substring in the error stack. We
haven't seen any errors in Sentry for the last 90 days, which is
suspicious. Let's verify if this condition is still valid and ensure we
aren't missing any false positives due to this logic.