Moved the version update to happen _after_ changelog
Previously it was causing issues for the changelog generator, since it
"couldn't find a common ancestor" when the version was updated and
commited first. Moving the version update to happen after resolves this,
and is PR'd to the same branch so that these changes happen at once.

I fear this might still lead to the same issue of common ancestor, but
we're merging into main, and then branching to the release branch, so
hopefully that doesn't happen?